C scripting interface, Hot-reloading, Improved reflection interface (#570)
* Initial test of concurrent Tecs transactions outside of DLL

* Start generating C api for component access

* Generate struct definitions based on reflected fields

* Proper C function calls through auto-generated api!

* Fix rust build

* Update Tecs

* Rename XRView.hh to XrView.hh (stupid Windows)

* Linux build fixes

* More shadowing fixes

* More typename

* Update Tecs c abi to use codegen function names

* Change reflection Access to AccessMut
Start cleaning up PhysicsUpdateLock

* Refactor script registration to use DynamicLock

* Update Tecs

* Adjust cmake dependencies

* Try and fix dependencies

* Add support for vectors through c abi and passing locks

* Update Tecs

* Implement script hot-reloading and camera_view in pure C

* Load scripts with cross-platform names

* Stop leaking metadata

* Stop leaking ScriptState::scriptData

* Refactor script loading to support dynamic libraries and better
loading/reloading

* Update script API for better thread safety (new/copy/free context funcs)

* Fixes for script state lifetime

* Move library loading to scene level and fix integration tests

* Fix dynamic script state reloading

* Set up library support instead of one script per dll

* Add signal and entity ref bindings, add new game_scripts library

* Fix undefined void * conversion behavior in function pointers

* Fix event polling being unsafe

* Fix metallic-roughness override and interactive gltf prefabs

* Create InlineString type, fix game of life

* Add union support and C-compatible EventData

* Codegen fixes and start making StructMetadata C-compatible

* Add reflection struct metadata support to C api

* Implement FlatSet and eliminate many heap allocations with
InlineVector/InlineString

* Refactor GuiRenderable layout to prepare for script defined guis

* Revert force constraint changes

* Set up GuiScript for defining ImGui windows

* Fix frame changes in tests

* Fix force-joints tests, allow Clang-format 20.1

* Update imgui and fix build on Clang 20

* Update tracy and fix clang/gcc build

* Fix screenshot diff compare

* Review cleanup

* Update tracy binaries, disable old formatter versions, add robin_hood
debug visualization definition

+++ include/strayphotons/game.h
@@ -20,7 +20,7 @@ extern "C" {
 typedef sp::CGameContext sp_game_t;
 #else
 typedef void sp_game_t;
-typedef unsigned int bool;
+typedef uint8_t bool;
 #endif
 
 struct GLFWwindow;
