+++ tests/unit/signal-caching.cc
@@ -32,7 +32,7 @@ namespace SignalCachingTests {
         if (it == nodes.end()) {
             Abortf("Could not find node in list: %s", node->text);
         } else {
-            AssertEqual(it - nodes.begin(), expectedIndex, "Unexpected node index");
+            AssertEqual((size_t)(it - nodes.begin()), expectedIndex, "Unexpected node index");
         }
     }
